Refactor SyncExecutor to SyncExecuteEngine
