+++ sharding-scaling/sharding-scaling-core/src/main/java/org/apache/shardingsphere/shardingscaling/core/execute/engine/SyncTaskExecuteEngine.java
@@ -35,5 +35,5 @@ public interface SyncTaskExecuteEngine {
      * @param syncRunners sync runner list
      * @return listenable future
      */
-    List<ListenableFuture> submit(List<SyncRunner> syncRunners);
+    List<ListenableFuture<Object>> submit(List<SyncRunner> syncRunners);
 }
